Clean rest session classes
