+++ config/default/restSessionFactory.conf.php
@@ -1,0 +1,7 @@
+<?php
+
+return new \oat\tao\model\session\restSessionFactory\RestSessionFactory(array(
+    \oat\tao\model\session\restSessionFactory\RestSessionFactory::OPTION_BUILDER => array(
+        \oat\tao\model\session\restSessionFactory\builder\HttpBasicAuthBuilder::class
+    )
+));
